Homework 01
Limiting upward and downward movement is achieved by constraints on the vertical coordinates of the spacecraft.

When moving upward, a check is made to see if the y-coordinate of the ship is less than 0 (the upper limit of the screen). If so, the y-coordinate is set to 0 to prevent the ship from moving upward off the screen.

When moving downward, the bottom edge of the ship (calculated as self.rect.y + self.rect.height) is checked to see if it is greater than the height of the screen (SCREEN_HEIGHT). If so, the y-coordinate is adjusted so that the ship stays within the lower limits of the screen and the Pac-Man "teleport" effect is performed so that the ship appears on the other side.